feat(web): collapse Review + Wi-Fi steps into a terminal Apply step
The setup wizard used to walk 5 steps:

  Home Overview → Layout → Wi-Fi → Security → Final Review

The user picked a Wi-Fi network at step 3, walked through Security
at step 4, and only at step 5 (Review) did the device actually try
to join the home network. The summary at step 5 read like an
informational screen but quietly hosted the destructive commit
behind a "Complete setup" button — the decision and the action
were three steps apart.

This change collapses the Wi-Fi credential collection and the Final
Review summary into a single terminal step (#597):

  Home Overview → Layout → Security → Apply

The new Apply step renders:

  - Top: the same summary list the Review step used to render
    (homeName, country, timezone, layout, wifi placeholder,
    admin-password marker).
  - Middle: the Wi-Fi network list + inline password (everything
    that used to live in `wifi-step.tsx`).
  - Bottom: a single "Save and switch network" CTA that opens the
    existing HandoffModal (#596) and triggers the commit + handoff
    overlay.

Wi-Fi gets last billing because the network switch is the
destructive action that ends the wizard. Everything above is
reversible from the same screen.

Changes:

- New `apps/web/src/features/setup/apply/apply-step.tsx` + barrel
  + smoke tests. Carries over the supervisor scan / commit
  helpers from the old wifi-step + review-step.
- Deleted `apps/web/src/features/setup/wifi/wifi-step.tsx` +
  `wifi-step.test.tsx` + `wifi/index.ts`. `handoff-modal.tsx`,
  `handoff-overlay.tsx`, `use-qr-code.ts`, `wifi-icons.tsx`
  survive under `wifi/` and are imported directly by apply-step
  (no barrel needed).
- Deleted the entire `apps/web/src/features/setup/review/`
  folder (no other consumer).
- `setup-route.tsx` — `WizardStepId` drops `'wifi'`, swaps
  `'review'` → `'apply'`; the SETUP_STEPS table reorders to
  the new 4-step shape. Step-icon paths drop the wifi/review
  entries and add `apply` (the old wifi-bars glyph, since the
  apply step's destructive action is the network switch).
- `setup-route.test.tsx` — assertions updated for the 4-step
  shape and the new "Save and switch network" CTA copy.
- i18n: full `setup.review.*` and `setup.wifi.*` (excluding the
  modal-scoped `setup.wifi.handoff.*`, which stays where it is)
  namespaces merge into `setup.apply.*` with sub-keys
  `setup.apply.summary.*` and `setup.apply.wifi.*`. EN + TR
  parity.
- E2E `setup-wizard.spec.ts` — walks 4 steps, clicks the
  embedded PreviewGuest card on the apply step, hits "Save and
  switch network" to commit.

44/44 setup tests green locally.

Closes #597
