+++ apps/web/src/features/setup/apply/index.ts
@@ -1,0 +1,1 @@
+export { ApplyStep } from './apply-step';
