--- apps/web/src/features/setup/review/index.ts
@@ -1,1 +1,0 @@
-export { ReviewStep } from './review-step';
