--- apps/web/src/features/setup/wifi/index.ts
@@ -1,1 +1,0 @@
-export { WifiStep } from './wifi-step';
